long -> int.

+++ src/main/java/de/eeriedaily/namqp/v08/Transmittable.java
@@ -7,7 +7,7 @@ import org.jboss.netty.buffer.ChannelBuffer;
  */
 public interface Transmittable {
 
-    public long getSize();
+    public int getSize();
 
     public void writeTo(ChannelBuffer channelBuffer);
